refactor runner functions
