+++ website/src/plugins/rspressPluginPlayground/cli/utils/getSourcesAndFiles.ts
@@ -8,7 +8,7 @@ import {
   transpileModule,
 } from 'typescript'
 import { transformAssetPaths } from './transformAssetPaths'
-import { APP_FILE_NAME } from '../../constants/files'
+import { APP_FILE_NAME } from '../../constants/stringConstants'
 
 const TRANSPILE_OPTIONS: TranspileOptions = {
   reportDiagnostics: false,
