--- website/src/plugins/rspressPluginPlayground/constants/babel.ts
@@ -1,2 +1,0 @@
-export const DEFAULT_BABEL_URL =
-  'https://cdn.jsdelivr.net/npm/@babel/standalone@7.26.4/babel.min.js'
