+++ website/src/plugins/rspressPluginPlayground/constants/babelTransform/index.ts
@@ -1,0 +1,1 @@
+export { babelTransform } from './babelTransform'
