--- website/src/plugins/rspressPluginPlayground/constants/files.ts
@@ -1,1 +1,0 @@
-export const APP_FILE_NAME = 'TS'
