+++ website/src/plugins/rspressPluginPlayground/constants/stringConstants.ts
@@ -1,0 +1,2 @@
+export const APP_FILE_NAME = 'TS'
+export const GET_IMPORT_FUNC_NAME = '__get_import'
