--- website/src/plugins/rspressPluginPlayground/web/Playground/Runner/constants.ts
@@ -1,1 +1,0 @@
-export const GET_IMPORT = '__get_import'
